moved DefinerResult and fixed DefinerUnitTest
